+++ resources/wa/index.d.ts
@@ -1,6 +1,7 @@
 declare const _default: {
     2017: any;
     2022: any;
+    2025: any;
 };
 export default _default;
 //# sourceMappingURL=index.d.ts.map