+++ resources/wa/index.js
@@ -1,7 +1,9 @@
 import edition2017 from './2017.js'
 import edition2022 from './2022.js'
+import edition2025 from './2025.js'
 
 export default {
 	2017: edition2017,
 	2022: edition2022,
+	2025: edition2025,
 }